i2c: Adds a context object for i2c
Rather than a custom I2C_DEVICE_DT define macro use include
ordering and hooks into the original DEVICE_ macros to create a
common i2c context object used as a container type for the
device state.

This eliminates the need for any i2c driver to understand how to use
the context, initialize the context, or allocate the context.

It may be opted out of by not including i2c_context as there is
a magic word placed in memory to enable runtime checks, when needed,
if the driver has opted in to using the common code and data.
