+++ drivers/i2c/i2c_bitbang.c
@@ -19,7 +19,7 @@
 
 #include <errno.h>
 #include <kernel.h>
-#include <drivers/i2c.h>
+#include "i2c_context.h"
 #include "i2c_bitbang.h"
 
 /*
