+++ drivers/i2c/i2c_npcx_controller.h
@@ -7,7 +7,7 @@
 #ifndef ZEPHYR_DRIVERS_I2C_I2C_NPCX_CONTROLLER_H_
 #define ZEPHYR_DRIVERS_I2C_I2C_NPCX_CONTROLLER_H_
 
-#include <device.h>
+#include <i2c_context.h>
 
 #ifdef __cplusplus
 extern "C" {
