+++ drivers/i2c/i2c_rcar.c
@@ -7,10 +7,10 @@
 #define DT_DRV_COMPAT renesas_rcar_i2c
 
 #include <errno.h>
-#include <device.h>
+#include "i2c_context.h"
 #include <devicetree.h>
 #include <soc.h>
-#include <drivers/i2c.h>
+#include "i2c_context.h"
 #include <drivers/clock_control.h>
 #include <drivers/clock_control/rcar_clock_control.h>
 
@@ -367,7 +367,7 @@ static const struct i2c_driver_api i2c_rcar_driver_api = {
 									       \
 	static struct i2c_rcar_data i2c_rcar_data_##n;			       \
 									       \
-	I2C_DEVICE_DT_INST_DEFINE(n,					       \
+	DEVICE_DT_INST_DEFINE(n,					       \
 			      i2c_rcar_init,				       \
 			      NULL,					       \
 			      &i2c_rcar_data_##n,			       \
