+++ drivers/i2c/i2c_sam_twi.c
@@ -18,10 +18,10 @@
 #include <sys/__assert.h>
 #include <stdbool.h>
 #include <kernel.h>
-#include <device.h>
+#include "i2c_context.h"
 #include <init.h>
 #include <soc.h>
-#include <drivers/i2c.h>
+#include "i2c_context.h"
 
 #define LOG_LEVEL CONFIG_I2C_LOG_LEVEL
 #include <logging/log.h>
@@ -354,7 +354,7 @@ static const struct i2c_driver_api i2c_sam_twi_driver_api = {
 									\
 	static struct i2c_sam_twi_dev_data i2c##n##_sam_data;		\
 									\
-	I2C_DEVICE_DT_INST_DEFINE(n, i2c_sam_twi_initialize,		\
+	DEVICE_DT_INST_DEFINE(n, i2c_sam_twi_initialize,		\
 			    NULL,					\
 			    &i2c##n##_sam_data, &i2c##n##_sam_config,	\
 			    POST_KERNEL, CONFIG_I2C_INIT_PRIORITY,	\
