+++ drivers/i2c/i2c_sbcon.c
@@ -14,9 +14,8 @@
  * hardware state of two-bit serial interfaces like I2C.
  */
 
-#include <device.h>
+#include "i2c_context.h"
 #include <errno.h>
-#include <drivers/i2c.h>
 #include "i2c_bitbang.h"
 
 /* SBCon hardware registers layout */
@@ -116,7 +115,7 @@ static const struct i2c_sbcon_config i2c_sbcon_dev_cfg_##_num = {	\
 	.sbcon		= (void *)DT_INST_REG_ADDR(_num), \
 };									\
 									\
-I2C_DEVICE_DT_INST_DEFINE(_num,						\
+DEVICE_DT_INST_DEFINE(_num,						\
 	    i2c_sbcon_init,						\
 	    NULL,							\
 	    &i2c_sbcon_dev_data_##_num,					\
