+++ drivers/i2c/i2c_test.c
@@ -10,7 +10,7 @@
  */
 
 #include <zephyr.h>
-#include <drivers/i2c.h>
+#include "i2c_context.h"
 
 #define DT_DRV_COMPAT vnd_i2c
 
@@ -38,7 +38,7 @@ static int vnd_i2c_init(const struct device *dev)
 }
 
 #define VND_I2C_INIT(n)						\
-	I2C_DEVICE_DT_INST_DEFINE(n, vnd_i2c_init, NULL,			\
+	DEVICE_DT_INST_DEFINE(n, vnd_i2c_init, NULL,			\
 			      NULL, NULL, POST_KERNEL,			\
 			      CONFIG_KERNEL_INIT_PRIORITY_DEVICE,	\
 			      &vnd_i2c_api);
