+++ include/drivers/i2c.h
@@ -192,6 +192,12 @@ struct i2c_slave_driver_api {
 	i2c_slave_api_unregister_t driver_unregister;
 };
 
+void z_i2c_xfer_stats(const struct device *dev, struct i2c_msg *msgs,
+			     uint8_t num_msgs);
+
+/* magic word value for detecting if common data is setup for driver */
+#define Z_I2C_MAGIC 0x1221EFFE
+
 /**
  * @endcond
  */
@@ -282,6 +288,7 @@ typedef int (*i2c_slave_read_requested_cb_t)(
 typedef int (*i2c_slave_read_processed_cb_t)(
 		struct i2c_slave_config *config, uint8_t *val);
 
+
 /** @brief Function called when a stop condition is observed after a
  * start condition addressed to a particular device.
  *
@@ -337,161 +344,9 @@ struct i2c_slave_config {
 	const struct i2c_slave_callbacks *callbacks;
 };
 
-#if defined(CONFIG_I2C_STATS) || defined(__DOXYGEN__)
-
-#include <stats/stats.h>
-
-/** @cond INTERNAL_HIDDEN */
-
-STATS_SECT_START(i2c)
-STATS_SECT_ENTRY32(bytes_read)
-STATS_SECT_ENTRY32(bytes_written)
-STATS_SECT_ENTRY32(message_count)
-STATS_SECT_ENTRY32(transfer_call_count)
-STATS_SECT_END;
-
-STATS_NAME_START(i2c)
-STATS_NAME(i2c, bytes_read)
-STATS_NAME(i2c, bytes_written)
-STATS_NAME(i2c, message_count)
-STATS_NAME(i2c, transfer_call_count)
-STATS_NAME_END(i2c);
-
-/** @endcond */
 
 
-/**
- * @brief I2C specific device state which allows for i2c device class specific additions
- */
-struct i2c_device_state {
-	struct device_state devstate;
-	struct stats_i2c stats;
-};
-
-/**
- * @brief Updates the i2c stats for i2c transfers
- *
- * @param dev I2C device to update stats for
- * @param msgs Array of struct i2c_msg
- * @param num_msgs Number of i2c_msgs
- */
-static inline void i2c_xfer_stats(const struct device *dev, struct i2c_msg *msgs,
-				  uint8_t num_msgs)
-{
-	struct i2c_device_state *state =
-		CONTAINER_OF(dev->state, struct i2c_device_state, devstate);
-	uint32_t bytes_read = 0U;
-	uint32_t bytes_written = 0U;
-
-	STATS_INC(state->stats, transfer_call_count);
-	STATS_INCN(state->stats, message_count, num_msgs);
-	for (uint8_t i = 0U; i < num_msgs; i++) {
-		if (msgs[i].flags & I2C_MSG_READ) {
-			bytes_read += msgs[i].len;
-		}
-		if (msgs[i].flags & I2C_MSG_WRITE) {
-			bytes_written += msgs[i].len;
-		}
-	}
-	STATS_INCN(state->stats, bytes_read, bytes_read);
-	STATS_INCN(state->stats, bytes_written, bytes_written);
-}
-
-/** @cond INTERNAL_HIDDEN */
-
-/**
- * @brief Define a statically allocated and section assigned i2c device state
- */
-#define Z_I2C_DEVICE_STATE_DEFINE(node_id, dev_name)	\
-	static struct i2c_device_state Z_DEVICE_STATE_NAME(dev_name)	\
-	__attribute__((__section__(".z_devstate")));
-
-/**
- * @brief Define an i2c device init wrapper function
- *
- * This does device instance specific initialization of common data (such as stats)
- * and calls the given init_fn
- */
-#define Z_I2C_INIT_FN(dev_name, init_fn)					\
-	static inline int UTIL_CAT(dev_name, _init)(const struct device *dev) \
-	{								\
-		struct i2c_device_state *state =			\
-			CONTAINER_OF(dev->state, struct i2c_device_state, devstate); \
-		stats_init(&state->stats.s_hdr, STATS_SIZE_32, 4,	\
-			   STATS_NAME_INIT_PARMS(i2c));			\
-		stats_register(dev->name, &(state->stats.s_hdr));	\
-		return init_fn(dev);					\
-	}
-
-/** @endcond */
-
-/**
- * @brief Like DEVICE_DT_DEFINE() with I2C specifics.
- *
- * @details Defines a device which implements the I2C API. May
- * generate a custom device_state container struct and init_fn
- * wrapper when needed depending on I2C @kconfig{CONFIG_I2C_STATS}.
- *
- * @param node_id The devicetree node identifier.
- *
- * @param init_fn Name of the init function of the driver.
- *
- * @param pm_device PM device resources reference (NULL if device does not use PM).
- *
- * @param data_ptr Pointer to the device's private data.
- *
- * @param cfg_ptr The address to the structure containing the
- * configuration information for this instance of the driver.
- *
- * @param level The initialization level. See SYS_INIT() for
- * details.
- *
- * @param prio Priority within the selected initialization level. See
- * SYS_INIT() for details.
- *
- * @param api_ptr Provides an initial pointer to the API function struct
- * used by the driver. Can be NULL.
- */
-#define I2C_DEVICE_DT_DEFINE(node_id, init_fn, pm_device,		\
-			     data_ptr, cfg_ptr, level, prio,		\
-			     api_ptr, ...)				\
-	Z_I2C_DEVICE_STATE_DEFINE(node_id, Z_DEVICE_DT_DEV_NAME(node_id)); \
-	Z_I2C_INIT_FN(Z_DEVICE_DT_DEV_NAME(node_id), init_fn)		\
-	Z_DEVICE_DEFINE(node_id, Z_DEVICE_DT_DEV_NAME(node_id),		\
-			DEVICE_DT_NAME(node_id),			\
-			&UTIL_CAT(Z_DEVICE_DT_DEV_NAME(node_id), _init), \
-			pm_device,					\
-			data_ptr, cfg_ptr, level, prio,			\
-			api_ptr,					\
-			&(Z_DEVICE_STATE_NAME(Z_DEVICE_DT_DEV_NAME(node_id)).devstate), \
-			__VA_ARGS__)
-
-#else /* CONFIG_I2C_STATS */
-
-static inline void i2c_xfer_stats(const struct device *dev, struct i2c_msg *msgs,
-				  uint8_t num_msgs)
-{
-}
-
-#define I2C_DEVICE_DT_DEFINE(node_id, init_fn, pm_device,		\
-			     data_ptr, cfg_ptr, level, prio,		\
-			     api_ptr, ...)				\
-	DEVICE_DT_DEFINE(node_id, &init_fn, pm_device,			\
-			     data_ptr, cfg_ptr, level, prio,		\
-			     api_ptr, __VA_ARGS__)
-
-#endif /* CONFIG_I2C_STATS */
-
-/**
- * @brief Like I2C_DEVICE_DT_DEFINE() for an instance of a DT_DRV_COMPAT compatible
- *
- * @param inst instance number. This is replaced by
- * <tt>DT_DRV_COMPAT(inst)</tt> in the call to I2C_DEVICE_DT_DEFINE().
- *
- * @param ... other parameters as expected by I2C_DEVICE_DT_DEFINE().
- */
-#define I2C_DEVICE_DT_INST_DEFINE(inst, ...)		\
-	I2C_DEVICE_DT_DEFINE(DT_DRV_INST(inst), __VA_ARGS__)
+#include <device.h>
 
 
 /**
@@ -588,7 +443,7 @@ static inline int z_impl_i2c_transfer(const struct device *dev,
 
 	int res =  api->transfer(dev, msgs, num_msgs, addr);
 
-	i2c_xfer_stats(dev, msgs, num_msgs);
+	z_i2c_xfer_stats(dev, msgs, num_msgs);
 
 	return res;
 }
